Break_LLM_CHAT_V2/requirements.txt
selenium>=4.0.0
webdriver-manager>=4.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
openai>=1.0.0
python-dotenv>=1.0.0
